fix: rename test-data to test_data folder
